+++ CHANGELOG.md
@@ -6,6 +6,12 @@ See also our [versioning policy](https://amici.readthedocs.io/en/latest/versioni
 
 ### v1.1.1 (unreleased)
 
+**Features**
+
+* The `fiddy` adapter (`amici.adapters.fiddy`) and the `fiddy` package itself
+  were redesigned, making finite-difference gradient checks much more
+  robust and requiring few, if any, hyperparameters.
+
 **Fixes**
 
 * There are no more reserved names: previously, model import or